4a8f88e48381b4c2ab341fa1eb931c12f40e3ee6 tests: Non-device tests do not need emulators compiled (Andrew Chow)
9b31a96f135381d10cc215173e9b6f17ad92035a tests: Only send kill signal if process isn't already dead (Andrew Chow)
9ab7373f076fb9e2a18d9b00c26a998e11ce8ca6 tests: Run each device's test suite individually (Andrew Chow)
3f82e60f558343ca95a772698235b60ad4dc4e23 test, trezor: report model T in tests correctly (Andrew Chow)
a741211819657eb0352d78dd6cbff49498243394 tests: move setUp and tearDown to DeviceTestCase (Andrew Chow)
e5808ac2766fbd3119a0cf0b1cb6871a4cafa3ba tests, ledger: Have speculos do button presses internally (Andrew Chow)
b8ba161df52769a34f9e183b9bc4d0e2945ec60b tests: use Travis' default xcode version (Andrew Chow)
4a4bf70fcb6ddb9a9b17e86f58569971151c7b96 tests: Use new speculos cli args (Andrew Chow)
cb138623786c52083cd69254dbd518a0de5661e4 tests: checkout Trezor T v2.2.0 (Andrew Chow)
83b60ad41c816464136f2f3db0f2a1aa352ad743 tests: coldcard simulator use --ms instead of -m (Andrew Chow)
cd0720fe0ddb7768e93ac9aaa940e80710fd66aa tests: Use p2sh-segwit address type for the p2sh-segwit import (Andrew Chow)
535505516f77caf6d9154ef0a40d1fbfc6a8b063 update Trezor 1 build script (Andrew Chow)
4b1c06f6d0aeb25c0b3c63786bce8fa107f61b0f Update keepkey build script (Andrew Chow)

Pull request description:

  For the past 2 weeks, travis has been failing because hardware wallet vendors have been updating their firmware (and emulators) in a way that was incompatible with the HWI tests. I've finally gotten around to fixing that.

  Things that were updated:
  * Keepkey build script. They switched to Python 3 and require a newer version of nanopb. Some other build commands had to be modified.
  * Trezor 1 build script. `script/setup` needs to be run in the pipenv.
  * Trezor T master branch incompatibility. Trezor T has made some major change that makes HWI incompatible with their current master branch (and probably future firmware versions too). For now, we checkout the 2.2.0 (latest as of this PR) firmware for testing.
  * Speculos command line arguments. Speculos changed their command line argument names.
  * Speculos patched to just do the key presses internally. Instead of having us to run another thread to send the correct key presses at the right time, we modify speculos to do that itself.
  * Coldcard command line arguments. Coldcard changed some command line argument names.
  * Some ismine thing is "broken" in Core, added a workaround for that.
  * Each devices tests are run with their own test runner. Global emulators will only be run for their device's tests.
  * MacOS build will use travis' defaul xcode version so that we don't run into issues with it being outdated and no longer supported by travis

Top commit has no ACKs.

Tree-SHA512: 0a5b6eb759a57f05254814d3c90e786989e41474d15b95393c08ef68e515fa31847129cef0933e0c801b7dea882abf3674438c0483ca57824024e25e6a566c45

+++ test/test_coldcard.py
@@ -5,6 +5,7 @@ import atexit
 import os
 import signal
 import subprocess
+import sys
 import time
 import unittest
 
@@ -13,7 +14,7 @@ from test_device import DeviceTestCase, start_bitcoind, TestDeviceConnect, TestD
 
 def coldcard_test_suite(simulator, rpc, userpass, interface):
     # Start the Coldcard simulator
-    coldcard_proc = subprocess.Popen(['python3', os.path.basename(simulator), '-m'], cwd=os.path.dirname(simulator), stdout=subprocess.DEVNULL, preexec_fn=os.setsid)
+    coldcard_proc = subprocess.Popen(['python3', os.path.basename(simulator), '--ms'], cwd=os.path.dirname(simulator), stdout=subprocess.DEVNULL, preexec_fn=os.setsid)
     # Wait for simulator to be up
     while True:
         try:
@@ -31,8 +32,9 @@ def coldcard_test_suite(simulator, rpc, userpass, interface):
     # Cleanup
 
     def cleanup_simulator():
-        os.killpg(os.getpgid(coldcard_proc.pid), signal.SIGTERM)
-        os.waitpid(os.getpgid(coldcard_proc.pid), 0)
+        if coldcard_proc.poll() is None:
+            os.killpg(os.getpgid(coldcard_proc.pid), signal.SIGTERM)
+            os.waitpid(os.getpgid(coldcard_proc.pid), 0)
     atexit.register(cleanup_simulator)
 
     # Coldcard specific management command tests
@@ -101,7 +103,11 @@ def coldcard_test_suite(simulator, rpc, userpass, interface):
     suite.addTest(DeviceTestCase.parameterize(TestDisplayAddress, rpc, userpass, 'coldcard', 'coldcard', '/tmp/ckcc-simulator.sock', '0f056943', 'tpubDDpWvmUrPZrhSPmUzCMBHffvC3HyMAPnWDSAQNBTnj1iZeJa7BZQEttFiP4DS4GCcXQHezdXhn86Hj6LHX5EDstXPWrMaSneRWM8yUf6NFd', interface=interface))
     suite.addTest(DeviceTestCase.parameterize(TestSignMessage, rpc, userpass, 'coldcard', 'coldcard', '/tmp/ckcc-simulator.sock', '0f056943', 'tpubDDpWvmUrPZrhSPmUzCMBHffvC3HyMAPnWDSAQNBTnj1iZeJa7BZQEttFiP4DS4GCcXQHezdXhn86Hj6LHX5EDstXPWrMaSneRWM8yUf6NFd', interface=interface))
     suite.addTest(DeviceTestCase.parameterize(TestSignTx, rpc, userpass, 'coldcard', 'coldcard', '/tmp/ckcc-simulator.sock', '0f056943', 'tpubDDpWvmUrPZrhSPmUzCMBHffvC3HyMAPnWDSAQNBTnj1iZeJa7BZQEttFiP4DS4GCcXQHezdXhn86Hj6LHX5EDstXPWrMaSneRWM8yUf6NFd', interface=interface))
-    return suite
+
+    result = unittest.TextTestRunner(stream=sys.stdout, verbosity=2).run(suite)
+    cleanup_simulator()
+    atexit.unregister(cleanup_simulator)
+    return result.wasSuccessful()
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Test Coldcard implementation')
@@ -113,5 +119,4 @@ if __name__ == '__main__':
     # Start bitcoind
     rpc, userpass = start_bitcoind(args.bitcoind)
 
-    suite = coldcard_test_suite(args.simulator, rpc, userpass, args.interface)
-    unittest.TextTestRunner(verbosity=2).run(suite)
+    sys.exit(not coldcard_test_suite(args.simulator, rpc, userpass, args.interface))
